@@ -3,8 +3,10 @@
   "description": "Find open issues from open source dependencies of your project.",
   "version": "1.5.0",
   "author": "Scott Davis <stdavis@utah.gov>",
+  "type": "module",
+  "exports": "./index.js",
   "bin": {
-    "good-samaritan": "./bin/run"
+    "good-samaritan": "./bin/run.js"
   },
   "bugs": "https://github.com/stdavis/good-samaritan/issues",
   "dependencies": {
@@ -23,18 +25,20 @@
     "read-pkg": "5.2.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.16.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
     "@oclif/test": "^1.2.8",
     "chai": "^4.3.0",
     "debug": "^4.3.1",
     "eslint": "^7.19.0",
     "eslint-config-prettier": "^7.2.0",
-    "jest": "^26.6.3",
+    "jest": "^27.4.7",
     "msw": "^0.26.2",
     "prettier": "^2.2.1",
     "standard-version": "^9.1.0"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=12.0.0"
   },
   "files": [
     "/bin",
@@ -52,7 +56,7 @@
   "repository": "stdavis/good-samaritan",
   "scripts": {
     "posttest": "eslint .",
-    "test": "jest --watchAll --coverage --silent",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watchAll --coverage --silent",
     "release": "standard-version"
   },
   "watch": {
