@@ -1,4 +1,5 @@
-const { getIssues, processIssues, getOctokitInstance } = require('./issues');
+import { getIssues, processIssues, getOctokitInstance } from './issues';
+import { jest } from '@jest/globals';
 
 describe('issues', () => {
   const largeIssueList = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17];
